==> Documents/1.Analysis/3.2_Impute_RRBS.Rmd <==
---
title: "DNA Methylation in the DO"
author: Anna L Tyler
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output: 
  html_document:
    code_folding: hide
    collapsed: no
    toc: yes
    toc_float: yes
---

The purpose of this workflow is to impute DNA methylation states
in DO mice and map local gene expression.

```{r load_libraries}
library(here)
is.interactive = FALSE
#is.interactive = TRUE

delete_previous = FALSE
#delete_previous = TRUE
```

```{r set_param}
females_only = FALSE
chow_only = TRUE
```


```{r source_code}
all.code.dir <- list.files(here("Code"), full.names = TRUE)
for(i in 1:length(all.code.dir)){
	all.fun <- list.files(all.code.dir[i], full.names = TRUE)
	for(j in 1:length(all.fun)){source(all.fun[j])}
}
```


```{r load_other_lib}
needed.packages <- c("qtl2", "abind", "pheatmap", "parallel", "doParallel")
load_libraries(needed.packages)
```


```{r read_data}
gene.info <- readRDS(here("Data", "RNASeq", "RNASeq_gene_info.RData"))
strain.table <- as.matrix(read.delim(here("Data", "support_files", "strain.color.table.txt"),
comment.char = "!", header = FALSE))

vars <- load(here("Data", "DOQTL", "Svenson_DO850_for_eQTL_viewer_v9.RData"))
expr <- dataset.mrna$data$rz
covar <- dataset.mrna$covar.matrix

col.table <- as.matrix(read.table(here("Data", "support_files", "strain.color.table.txt"), 
sep = "", comment.char = "%", stringsAsFactors = FALSE))
strains <- col.table[,6]
```

Read in the gene methylation data from 3.1.1_RRBS_Analysis.Rmd

```{r gene_methyl}
methyl.mat.file <- here("Results", "RRBS", "Aligned.Methyl.Mats.RDS")
methyl.mats <- readRDS(methyl.mat.file)

#transcripts that are shared between RRBS and DO expression data
methyl.id <- gene.info[match(names(methyl.mats), gene.info[,"external_gene_name"]),"ensembl_gene_id"]
methyl.id <- methyl.id[which(!is.na(methyl.id))]
ordered.transcript.info <- gene.info[match(methyl.id, gene.info[,"ensembl_gene_id"]),]
```

```{r fun}
get_nearest_marker <- function(chr, start.pos){
    if(is.na(chr)){return(NA)}
    if(chr == "MT"){return(NA)}
    chr.locale <- which(names(map) == chr)
    nearest.marker <- get.nearest.pt(map[[chr.locale]], start.pos/1e6)
    marker.geno <- genoprobs[[chr.locale]][,,nearest.marker]
    return(marker.geno)
}

rrbs_geno <- function(gene.rrbs.array, haplotype.matrix, gene.chr, perm.order = NULL){
    if(length(gene.rrbs.array) == 1){return(NA)}
    if(length(haplotype.matrix) <= 1){return(NA)}
    #filter out NAs
    has.all <- which(apply(gene.rrbs.array, 2, function(x) all(!is.na(x))))
    if(length(has.all) <= 1){return(NA)}
    #plot.methyl.mat(gene.rrbs.array[,has.all])
    #boxplot(gene.rrbs.array[,has.all])
    #pheatmap(cor(gene.rrbs.array[,has.all]), cluster_rows = FALSE, cluster_cols = FALSE)

    #bin the methylation into 0, 50, and 100
    #this prevents transferring haplotypes directly
    #onto very small variations in methylation
    binned.rrbs <- t(apply(gene.rrbs.array[,has.all,drop=FALSE], 1, function(x) bin.vector(x, c(0, 50, 100))))
        
    #do the imputation, with a shuffled haplotype order if requested
    if(is.null(perm.order)){
      all.rrbs <- haplotype.matrix %*% binned.rrbs
    }else{
      perm.rrbs <- binned.rrbs[perm.order,]
      rownames(perm.rrbs) <- rownames(binned.rrbs)
      all.rrbs <- haplotype.matrix %*% perm.rrbs
    }
    #pheatmap(all.rrbs, cluster_cols = FALSE)
    #boxplot(all.rrbs) 
    #pheatmap(haplotype.matrix)
    #hap.cor <- matrix(NA, nrow = ncol(haplotype.matrix), ncol = ncol(all.rrbs))
    #rownames(hap.cor) <- colnames(haplotype.matrix)
    #colnames(hap.cor) <- colnames(all.rrbs)
    #for(i in 1:ncol(haplotype.matrix)){    
    #  hap.cor[i,] <- apply(all.rrbs, 2, function(x) cor(x, haplotype.matrix[,i]))
    #}
    #pheatmap(hap.cor, cluster_rows = FALSE, cluster_cols = FALSE)
    #which(hap.cor[1,] == 1)
    #gene.rrbs.array[,"67245515"]
    
    all.rrbs.mat <- abind(all.rrbs, round(100-all.rrbs), along = 3)
    dimnames(all.rrbs.mat)[[3]] <- c("A", "B")
    all.rrbs.array <- aperm(all.rrbs.mat, c(1,3,2))
    one.geno <- list(all.rrbs.array)

    #add attributes
    names(one.geno) <- gene.chr
    if(gene.chr == "X"){
      attr(one.geno, "is_x_chr") <- TRUE
    }else{
      attr(one.geno, "is_x_chr") <- FALSE
    }
    attr(one.geno, "crosstype") <- "do"
    attr(one.geno, "alleles") <- dimnames(all.rrbs.array)[[3]]
    attr(one.geno, "alleleprobs") <- TRUE
    attr(one.geno, "class") <- c("calc_genoprob", "list")
    return(one.geno)
}

```


## Get Genotypes for Nearest Markers

For every gene in our list, pull out the genotype matrix for the marker
nearest the TSS.

```{r set_gene}
gene.marker <- vector(mode = "list", length = length(methyl.id))
names(gene.marker) <- methyl.id
for(i in 1:length(methyl.id)){
  if(is.interactive){report.progress(i, length(methyl.id))}
  gene.chr <- ordered.transcript.info[i,"chromosome_name"]
  gene.pos <- ordered.transcript.info[i, "start_position"]
  if(gene.chr != "Y" && gene.chr != "MT"){
    gene.marker[[i]] <- get_nearest_marker(gene.chr,gene.pos) 
  }
}
```

## Impute DO methylation percent and map

Derive DNA methylation percentages for each mouse given 
the genoprobs and the methylation data compiled from 3.1.1_RRBS_Analysis.Rmd. 
This array will have DNA methylation percentage estimates for each 
individual by position.

To generate this array we multiply each DNA methylation matrix along 
the genome by the haplotype matrix for the gene.

```{r scan_rrbs_geno}
#convert to RRBS arrays and scan.
#The arrays are too big to store, so we
#scan them right away and only keep the 
#LOD scores.
strain.order <- order.strains(dimnames(genoprobs[[1]])[[2]], rownames(methyl.mats[[1]]), col.table)
rrbs.r2.file <- here("Results", "RRBS", "RRBS.R2.RDS")
rrbs.coef.file <- here("Results", "RRBS", "RRBS.coef.RDS")

if(females_only){
  rrbs.r2.file <- gsub(".RDS", "_females.RDS", rrbs.r2.file)
  rrbs.coef.file <- gsub(".RDS", "_females.RDS", rrbs.coef.file)

}
if(chow_only){
  rrbs.r2.file <- gsub(".RDS", "_chow.RDS", rrbs.r2.file)
  rrbs.coef.file <- gsub(".RDS", "_chow.RDS", rrbs.coef.file)
}

if(!file.exists(rrbs.r2.file) || delete_previous){
    sink(here("Results", "RRBS", "progress.txt"))
    all.rrbs.r2 <- all.rrbs.coef <- vector(mode = "list", length = length(methyl.id))
    names(all.rrbs.r2) <- names(all.rrbs.coef) <- methyl.id
    for(i in 1:length(all.rrbs.r2)){
        cat(i, ":")
        methyl.mat.locale <- which(names(methyl.mats) == ordered.transcript.info[i,"external_gene_name"])
        if(length(methyl.mats[[methyl.mat.locale]]) <= 1 || ncol(methyl.mats[[methyl.mat.locale]]) >= 2000){cat("No DNA methylation matrix\n");next()}

        transcript.chr <- ordered.transcript.info[i,"chromosome_name"]
        if(is.na(transcript.chr) || transcript.chr == "X" || transcript.chr == "MT"){cat(" Not on an autosome\n");next()}

        rrbs.mat <- methyl.mats[[methyl.mat.locale]][strain.order,,drop=FALSE]
        #replace NAs with 0s becuase missing CpG sites will not be methylated
        rrbs.mat[which(is.na(rrbs.mat))] <- 0
        hap.mat <- gene.marker[[i]]
        rrbs.array <- rrbs_geno(gene.rrbs.array = rrbs.mat, 
          haplotype.matrix = hap.mat, 
          gene.chr = ordered.transcript.info[i,"chromosome_name"])

        if(is.na(rrbs.array)){cat(" Cannot create RRBS array\n");next()}

        gene.expr <- expr[,which(colnames(expr) == methyl.id[i]),drop=FALSE]
        
        if(length(gene.expr) == 0){cat(" No gene expression\n");next()}

        sub.covar <- covar #duplicate covar in case we want to subset by both sex and diet
        if(females_only){
          sex.idx <- which(covar[,"sexM"] == 0)
        }else{
          sex.idx <- 1:nrow(covar)
        }
        if(chow_only){
          diet.idx <- which(covar[,"diethf"] == 0)
        }else{
          diet.idx <- 1:nrow(covar)
        }
        sub.idx <- intersect(sex.idx, diet.idx)

        adj.expr <- adjust(gene.expr[sub.idx,,drop=FALSE], covar[sub.idx,])
        
        common.ind <- intersect(rownames(adj.expr), rownames(genoprobs[[1]]))
        common.expr.locale <- match(common.ind, rownames(adj.expr))
        common.genoprob.locale <- match(common.ind, rownames(genoprobs[[1]]))

        regression.models <- apply(rrbs.array[[1]][common.genoprob.locale,1,], 2, function(x) if(var(x) > 0){lm(adj.expr[common.expr.locale,]~x)}else{NA})
        rrbs.r2 <- sapply(regression.models, function(x) if(length(x) > 1){summary(x)$adj.r.squared}else{NA})
        all.rrbs.r2[[i]] <- rrbs.r2  

        rrbs.coef <- sapply(regression.models, function(x) if(length(x) > 1){coef(x)[2]}else{NA})
        all.rrbs.coef[[i]] <- rrbs.coef
        
        cat("\n")
        #quartz();barplot(all.rrbs.r2[[i]])
    }
    sink()
    saveRDS(all.rrbs.r2, rrbs.r2.file)
    saveRDS(all.rrbs.coef, rrbs.coef.file)
}else{
  all.rrbs.r2 <- readRDS(rrbs.r2.file)
  all.rrbs.coef <- readRDS(rrbs.coef.file)
}
```

## Position Effects

As we did with the inbred expression, look for 
position-based effects in the outbred expression.

```{r center_coef}
gene.names <- gene.info[match(names(all.rrbs.coef), gene.info[,"ensembl_gene_id"]),"external_gene_name"]
centered.coord <- lapply(1:length(gene.names),
  function(x) if(length(all.rrbs.coef[[x]]) > 1){center.on.feature(gene.names[x], 
  gene.info, as.numeric(gsub(".x", "", names(all.rrbs.coef[[x]]))), 
  feature = "full")}else{NA})

ymin <- min(unlist(all.rrbs.coef), na.rm = TRUE)
ymax <- max(unlist(all.rrbs.coef), na.rm = TRUE)

nbins = 15
coord.bins <- segment.region(-0.5, 1.5, nbins)
plot.coord <- coord.bins[1:(length(coord.bins)-1)]
all.coord <- unlist(lapply(centered.coord, function(x) as.numeric(names(x))))
all.coef <- unlist(all.rrbs.coef)

binned.coef <- lapply(1:(length(coord.bins)-1), 
  function(x) all.coef[intersect(which(all.coord >= coord.bins[x]), 
  which(all.coord < coord.bins[(x+1)]))])

coef.mean <- sapply(binned.coef, function(x) mean(x, na.rm = TRUE))
coef.se <- sapply(binned.coef, function(x) sd(x, na.rm = TRUE)/sqrt(length(which(!is.na(x)))))
coef.conf <- sapply(binned.coef, function(x) t.test(x, conf.int = 99)$conf.int)

#ylim = c(min(coef.conf), max(coef.conf))
#plot(1:length(binned.coef), coef.mean, type = "l", ylim = ylim)
#points(1:14, coef.conf[1,], col = "red", type = "l")
#points(1:14, coef.conf[2,], col = "red", type = "l")
#abline(h = 0)

imp.rrbs.effect <- rbind(coef.mean, coef.conf)
colnames(imp.rrbs.effect) <- plot.coord

imputed.effect.file <- here("Results", "RRBS", "Imputed_RRBS_effect.RDS")
if(females_only){
  imputed.effect.file <- gsub(".RDS", "_females.RDS", imputed.effect.file)
}
if(chow_only){
  imputed.effect.file <- gsub(".RDS", "_chow.RDS", imputed.effect.file)
}
saveRDS(imp.rrbs.effect, imputed.effect.file)

plot.new()
plot.window(xlim = c(-0.5, 1.5), ylim = c(min(coef.mean-coef.se), max(coef.mean+coef.se)))
plot.poly.xy(poly.top.x = plot.coord, poly.top.y = coef.mean+coef.se, 
  poly.bottom.x = plot.coord, poly.bottom.y = coef.mean-coef.se,
  border = NULL, col = "gray", lwd = 1, new.plot = FALSE)
abline(h = 0)
axis(1)
axis(2)
abline(v = c(0, 1), col = "gray", lty = 2, lwd = 2)
mtext("Mean DNA Methylation Effect on Transcription in DO", side = 3, line = 1.5)

```

```{r centered_r2}
ymin <- min(unlist(all.rrbs.r2), na.rm = TRUE)
ymax <- max(unlist(all.rrbs.r2), na.rm = TRUE)

all.r2 <- unlist(all.rrbs.r2)

binned.r2 <- lapply(1:(length(coord.bins)-1), 
  function(x) all.r2[intersect(which(all.coord >= coord.bins[x]), 
  which(all.coord < coord.bins[(x+1)]))])

r2.mean <- sapply(binned.r2, function(x) mean(x, na.rm = TRUE))
r2.se <- sapply(binned.r2, function(x) sd(x, na.rm = TRUE)/sqrt(length(which(!is.na(x)))))

plot.new()
plot.window(xlim = c(-0.5, 1.5), ylim = c(min(r2.mean-r2.se), max(r2.mean+r2.se)))
plot.poly.xy(poly.top.x = plot.coord, poly.top.y = r2.mean+r2.se, 
  poly.bottom.x = plot.coord, poly.bottom.y = r2.mean-r2.se,
  border = NULL, col = "gray", lwd = 1, new.plot = FALSE)
abline(h = 0)
axis(1)
axis(2)
abline(v = c(0, 1), col = "gray", lty = 2, lwd = 2)
mtext("Mean DNA Methylation Variance Explained in DO", side = 3, line = 1.5)
```

## Permutations for RRBS

From 3.2_Impute_RRBS.Rmd

## Permute RRBS

We permuted the DNA methylation as we did for chromatin 
states. We permute the haplotype that is assigned to 
each DNA methylation vector, re-do the imputation, 
and recalculate the LOD scores and R2 values.


```{r perm_fun}

#This function and the perm data will be exported to the
#clusters
scan_one_rrbs <- function(perm.data, p){
    
  perm.order <- order.mat[p,]

  imp.geno <- rrbs_geno(gene.rrbs.array = perm.data$rrbs.mat, 
            haplotype.matrix = perm.data$hap.mat, 
            gene.chr = perm.data$chr, perm.order)
          
  if(is.na(imp.geno)){return(NA)}

  one.lod <- scan1(imp.geno, perm.data$gene.expr)
  #plot(as.numeric(rownames(all.rrbs.lod[[i]])), all.rrbs.lod[[i]][,1], type  = "h",xlab = "Genomic Position", ylab = "DNA Methylation LOD score")

  common.ind <- intersect(rownames(perm.data$gene.expr), rownames(imp.geno[[1]]))
  common.expr.locale <- match(common.ind, rownames(perm.data$gene.expr))
  common.genoprob.locale <- match(common.ind, rownames(imp.geno[[1]]))

  regression.models <- apply(imp.geno[[1]][common.genoprob.locale,1,], 2, 
      function(x) if(var(x) > 0){lm(perm.data$gene.expr[common.expr.locale,]~x)}else{NA})
  one.rrbs.r2 <- sapply(regression.models, function(x) if(length(x) > 1){summary(x)$adj.r.squared}else{NA})

  results <- list("one.lod" = one.lod, 
    "one.r2" = one.rrbs.r2)

  return(results)
}
```

```{r select_nperm}
nperm = 1000
```


```{r set_perm}
#select a set of permutations for which none of 
#the haplotypes remain in their original position.
#these are called derrangements
perm.mat <- derrangements(8)
#sample a random set of derraangements to use
#for all transcripts
perm.rows <- sample(1:nrow(perm.mat), nperm)
order.mat <- perm.mat[perm.rows,]
```



```{r perm_files}

perm.lod.rrbs.file <- here("Results", "RRBS", "Perm.RRBS.LOD.RDS")
perm.r2.rrbs.file <- here("Results", "RRBS", "Perm.RRBS.R2.RDS")

if(females_only){
  perm.lod.rrbs.file <- gsub(".RDS", "_females.RDS", perm.lod.rrbs.file)
  perm.r2.rrbs.file <- gsub(".RDS", "_females.RDS", perm.r2.rrbs.file)
}
if(chow_only){
  perm.lod.rrbs.file <- gsub(".RDS", "_chow.RDS", perm.lod.rrbs.file)
  perm.r2.rrbs.file <- gsub(".RDS", "_chow.RDS", perm.r2.rrbs.file)
}

if(file.exists(perm.lod.rrbs.file)){
  perm.rrbs.lod <- readRDS(perm.lod.rrbs.file)
  perm.rrbs.r2 <- readRDS(perm.r2.rrbs.file)

  #figure out where we left off.
  not.na <- which(!is.na(perm.rrbs.lod[,1]))  
  left.off <- max(not.na)
  start.at <- left.off + 1
}else{
  start.at <- 1
  #if we are at the beginning, intialize all the objects for holding
  #results. 
  perm.rrbs.lod <- perm.rrbs.r2 <- matrix(NA, nrow = length(methyl.id), ncol = nperm)
  rownames(perm.rrbs.lod) <- rownames(perm.rrbs.r2) <- methyl.id
}

```


```{r set_cl, warning = FALSE}
save.every = 100
if(start.at < length(methyl.id)){
  #set up the clusters if we need to run perutations
  n.cores <- detectCores() - 1
  to.export <- c("scan1", "scan1coef", "scan_one_rrbs", "rrbs_geno", 
    "order.mat", "bin.vector", "abind", "common.genoprob.locale", 
    "common.expr.locale")

  #create and register multiple cores for parallele permutations
  cl <- makeCluster(n.cores)
  registerDoParallel(cl)
  #export everything we need to the cluster
  #this takes time upfront, but the parallelization
  #saves a huge amount of time overall
  clusterExport(cl, to.export) 
}
```


```{r perm_rrbs}

if(start.at < length(methyl.id)){
  sink(here("Results", "RRBS", "progress.txt"))
  for(tr in start.at:length(perm.rrbs.lod)){
      cat(paste0(tr, ":"))
      methyl.mat.locale <- which(names(methyl.mats) == ordered.transcript.info[tr,"external_gene_name"])

      #checks for methylation matrix
      if(length(methyl.mats[[methyl.mat.locale]]) <= 1){cat(" Cannot locate array\n");next()} #it exists
      if(ncol(methyl.mats[[methyl.mat.locale]]) > 2000){cat(" Array too large\n");next()} #is is of a reasonable size

      transcript.chr <- ordered.transcript.info[tr,"chromosome_name"]

      #checks for chromosome
      if(is.na(transcript.chr) || transcript.chr == "X" || transcript.chr == "MT"){cat(" Transcript not on autosome\n");next()}
      
      rrbs.mat <- methyl.mats[[methyl.mat.locale]][strain.order,,drop=FALSE]
      #change missing CpG sites to 0 because if they are not a CpG site, they will not be methylated.
      rrbs.mat[which(is.na(rrbs.mat))] <- 0
      hap.mat <- gene.marker[[tr]]
      rrbs.array <- rrbs_geno(gene.rrbs.array = rrbs.mat, 
        haplotype.matrix = hap.mat, 
        gene.chr = ordered.transcript.info[i,"chromosome_name"])

      #check for rrbs.array
      if(is.na(rrbs.array)){cat(" No RRBS array\n"); next()}

      expr.locale <- which(colnames(expr) == methyl.id[tr])

      gene.expr.choice <- as.logical(length(expr.locale))
      
      #check for gene expr
      if(length(expr.locale) == 0){cat(" Cannot find expression\n");next()}
      gene.expr <- expr[,expr.locale,drop=FALSE]
      adj.expr <- adjust(gene.expr, covar)

      sub.covar <- covar #duplicate covar in case we want to subset by both sex and diet

        if(females_only){
          sex.idx <- which(covar[,"sexM"] == 0)
        }else{
          sex.idx <- 1:nrow(covar)
        }
        if(chow_only){
          diet.idx <- which(covar[,"diethf"] == 0)
        }else{
          diet.idx <- 1:nrow(covar)
        }
        sub.idx <- intersect(sex.idx, diet.idx)

      perm.data <- list("gene.expr" = gene.expr[sub.idx,,drop=FALSE], 
        "rrbs.mat" = rrbs.mat, "hap.mat" = hap.mat,
        "chr" = ordered.transcript.info[tr,"chromosome_name"])

      #send the new gene ID to the clusters
      clusterExport(cl, "perm.data")

      #do the permutations in parallel
      perm.rrbs.results <- parLapply(cl, 1:nperm, 
          function(p) scan_one_rrbs(perm.data, 
          p = p))

      #check for data returned from permutations
      if(!all(sapply(perm.rrbs.results, length) > 1)){cat(" No data returned\n");next()}
      
      perm.tr.lod <- lapply(perm.rrbs.results, function(x) x$one.lod)
      perm.tr.r2 <- lapply(perm.rrbs.results, function(x) x$one.r2)

      #keep the maximum r2 and LOD values from each permutation.
      perm.rrbs.lod[tr,] <- t(sapply(perm.tr.lod, function(x) max(x[,1])))
      perm.rrbs.r2[tr,] <- t(sapply(perm.tr.r2, function(x) max(x, na.rm = TRUE)))

      if(tr %% save.every == 0){
          saveRDS(perm.rrbs.lod, perm.lod.rrbs.file)
          saveRDS(perm.rrbs.r2, perm.r2.rrbs.file)
      }
    cat(" Finished permutations\n")
  } #end looping over transcripts
  #save again at the very end
  saveRDS(perm.rrbs.lod, perm.lod.rrbs.file)
  saveRDS(perm.rrbs.r2, perm.r2.rrbs.file)

  sink()

  #close the clusters at the end and save the final objects
  stopCluster(cl)

} #end case for needing to do more permutations.
```

## Assess permutations

To assess the permutations, we look at the distribution
of empirical p values. If this distribution is uniform,
then it is the imputation itself that is explaining the
LOD scores we see. If the distribution is skewed toward
0, then the DNA methylation is adding something beyond
the imputation itself.

The plots below show the distributions of p values for both 
the R2 statistic and the LOD score statistic. Both are
skewed toward 0. The dot plot shows that the p values for 
these two statistics are highly correlated, but not identical.

```{r assess_lod_perm, fig.width = 4, fig.height = 4}
geno.lod <- readRDS(here("Results", "ChromHMM", "14_states_C", "eQTL.lod.RDS"))
low.lod <- which(geno.lod[,"Additive"] <= 6)
low.lod.idx <- match(rownames(geno.lod)[low.lod], names(all.rrbs.r2))

max.rrbs.r2 <- sapply(all.rrbs.r2, function(x) if(!all(is.na(x))){max(x, na.rm = TRUE)}else{NA})
r2.p <- sapply(1:length(all.rrbs.r2), 
  function(x) if(!all(is.na(perm.rrbs.r2[x,]))){emp.p(max.rrbs.r2[x], perm.rrbs.r2[x,])}else{NA})
names(r2.p) <- names(all.rrbs.r2)

results.table <- cbind(max.rrbs.r2, r2.p)
perm.results.file <- here("Results", "RRBS", "Perm_results.RDS")

if(females_only){
  perm.results.file <- gsub(".RDS", "_females.RDS", perm.results.file)
}
if(chow_only){
  perm.results.file <- gsub(".RDS", "_chow.RDS", perm.results.file)
}

saveRDS(results.table, perm.results.file)

r2.p[low.lod.idx] <- NA
hist(r2.p, breaks = 100, main = "R2 p",
  xlab = "p value")

```

The following plots show each test statistic vs. its
p values. In general higher test statistics have lower
p values, but this is not strictly true. 


```{r stat_v_p, fig.width = 4, fig.height = 4}

plot(max.rrbs.r2, -log10(r2.p+1e-3), xlab = "Variance Explained", 
  ylab = "-log10(p)", main = "R2 vs. R2-log10(p)")
```

```{r example, eval = FALSE}
examples <- which(r2.p == 0)
i <- examples[5]
gene.name = gene.info[which(gene.info[,"ensembl_gene_id"] == names(r2.p)[i]),"external_gene_name"][1]
xlim <- c(min(c(perm.rrbs.r2[i,], max.rrbs.r2[i])), max(c(perm.rrbs.r2[i,], max.rrbs.r2[i])))
hist_with_points(c(perm.rrbs.r2[i,], max.rrbs.r2[i]), 
  col = c(rep("gray", ncol(perm.rrbs.r2)), "red"), xlab = "LOD", main = gene.name)
abline(v = max.rrbs.r2[i], col = "red")
```
