==> Documents/1.Analysis/5.0_Compare_Perm.Rmd <==
---
title: "5.0 Compare Permutations"
author: Anna L Tyler
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output: 
  html_document:
    code_folding: hide
    collapsed: no
---


## Introduction
The purpose of this workflow is to compare the permuted
effects of the different genomic features: chromatin state,
SNPs, and DNA methylation.

```{r libraries}
library(here)
num.states = 14
```

```{r source_code}
all.code.dir <- list.files(here("Code"), full.names = TRUE)
for(i in 1:length(all.code.dir)){
	all.fun <- list.files(all.code.dir[i], full.names = TRUE)
	for(j in 1:length(all.fun)){source(all.fun[j])}
}
```

```{r load_other_lib}
needed.packages <- c("gprofiler2", "pheatmap")
load_libraries(needed.packages)
```


```{r read_data}
gene.info <- read.csv(here("Data", "DOQTL", "mRNA_Info.csv"))

chrom.results <- readRDS(here("Results", "ChromHMM", paste0(num.states, "_states_C"), "Perm_results.RDS")) 
rrbs.results <- readRDS(here("Results", "RRBS", "Perm_results.RDS"))
snp.results <- readRDS(here("Results", "SNPs", "Perm_results.RDS"))

all.results <- list("Chromatin" = chrom.results, "DNA Methylation" = rrbs.results,
    "SNPs" = snp.results)
common.transcripts <- Reduce("intersect", lapply(all.results, rownames))
matched.idx <- lapply(all.results, 
    function(x) match(common.transcripts, rownames(x)))
feature.pairs <- pair.matrix(1:length(all.results))

matched.results <- lapply(1:length(all.results), 
    function(x) all.results[[x]][matched.idx[[x]],])
names(matched.results) <- names(all.results)
saveRDS(matched.results, here("Results", "General", "Permutation_P_values.RDS"))
```

The following plots show the p value distributions for each
genomic feature.

```{r perm_dist, warning = FALSE, fig.width = 9, fig.height = 3}
par(mfrow = c(1,3))
unif.p <- rep(NA, length(matched.results))
for(i in 1:length(matched.results)){
    hist(matched.results[[i]][,"r2.p"], breaks = 100, 
        main = names(matched.results)[i],
        xlab = "p value")
    unif.p[i] <- ks.test(matched.results[[i]][,"r2.p"], "punif")$p.value
}

```

```{r alpha}
alpha = 1/1000
```

The following Venn diagram shows the number of 
transcripts that had significant (p = `r alpha`) 
effects from each feature according to the permutations.

DNA methylation actually has the largest number of 
significant transcripts. And there is very little
overlap among the groups.

```{r sig_idx, warning = FALSE, message = FALSE}
sig.idx <- lapply(matched.results, function(x) which(x[,2] <= alpha))
plotVenn(sig.idx)

all.intersections <- lapply(1:nrow(feature.pairs), 
    function(x) common.transcripts[intersect(sig.idx[[feature.pairs[x,1]]], sig.idx[[feature.pairs[x,2]]])])
names(all.intersections) <- apply(feature.pairs, 1, 
    function(x) paste0(names(matched.results)[x[1]], "_", names(matched.results)[x[2]]))

#intersected.effects <- lapply(matched.results, function(x) x[match(unlist(all.intersections), rownames(x)),1])
#boxplot(intersected.effects)
```

```{r eval = FALSE}
#run Figures.Rmd to load data and code and then check the following

plot_together("ENSMUSG00000007029")

```

The following plot shows enrichment terms for genes with 
significant effects of each imputed genomic feature.

```{r enrich, fig.width = 7, fig.height = 7}
sig.enrich <- lapply(sig.idx, function(x) gost(names(x), organism = "mmusculus",
    sources = c("GO", "KEGG", "REACTOME")))

plot.enrichment.group(sig.enrich, transformation = function(x) log(x+1e-6), n.terms = 30)
```

The following boxplot shows the effect sizes of the significant effects
for each genomic feature.

```{r lod_comp}
sig.effect <- lapply(1:length(sig.idx), function(x) matched.results[[x]][sig.idx[[x]],1])
names(sig.effect) <- names(matched.results)
boxplot(sig.effect, main = "LOD scores of significant effects")
```


The following set of plots compare the variance explained by
the different genomic features 

```{r compare, fig.width = 8, fig.height = 8}
par(mfrow = c(2,2))
for(i in 1:nrow(feature.pairs)){
    pair1 <- feature.pairs[i,1]
    pair2 <- feature.pairs[i,2]
    plot(all.results[[pair1]][matched.idx[[pair1]],3], 
        all.results[[pair2]][matched.idx[[pair2]],3], 
        xlab = names(all.results)[pair1], ylab = names(all.results)[pair2])
    abline(0,1,col = "red")
}
```
